analyze lib
